Fixed backslash duplication for PostgreSQL

+++ lib/mshoplib/tests/MShop/Index/Manager/MySQLTest.php
@@ -22,23 +22,28 @@ class MySQLTest extends \PHPUnit_Framework_TestCase
 	public static function setUpBeforeClass()
 	{
 		$context = clone \TestHelperMShop::getContext();
-		$context->getConfig()->set( 'mshop/index/manager/text/name', 'MySQL' );
+		$config = $context->getConfig();
+		$dbadapter = $config->get( 'resource/db-index/adapter', $config->get( 'resource/db/adapter' ) );
 
-		$manager = new \Aimeos\MShop\Index\Manager\MySQL( $context );
-		$productManager = \Aimeos\MShop\Product\Manager\Factory::createManager( $context );
-
-		$search = $productManager->createSearch();
-		$conditions = array(
-			$search->compare( '==', 'product.code', array( 'CNC', 'CNE' ) ),
-			$search->compare( '==', 'product.editor', $context->getEditor() )
-		);
-		$search->setConditions( $search->combine( '&&', $conditions ) );
-		$result = $productManager->searchItems( $search, array( 'attribute', 'price', 'text', 'product' ) );
-
-		foreach( $result as $item )
+		if( $dbadapter === 'mysql' )
 		{
-			$manager->deleteItem( $item->getId() );
-			$manager->saveItem( $item );
+			$context->getConfig()->set( 'mshop/index/manager/text/name', 'MySQL' );
+			$manager = new \Aimeos\MShop\Index\Manager\MySQL( $context );
+			$productManager = \Aimeos\MShop\Product\Manager\Factory::createManager( $context );
+
+			$search = $productManager->createSearch();
+			$conditions = array(
+				$search->compare( '==', 'product.code', array( 'CNC', 'CNE' ) ),
+				$search->compare( '==', 'product.editor', $context->getEditor() )
+			);
+			$search->setConditions( $search->combine( '&&', $conditions ) );
+			$result = $productManager->searchItems( $search, array( 'attribute', 'price', 'text', 'product' ) );
+
+			foreach( $result as $item )
+			{
+				$manager->deleteItem( $item->getId() );
+				$manager->saveItem( $item );
+			}
 		}
 	}
 
@@ -46,17 +51,16 @@ class MySQLTest extends \PHPUnit_Framework_TestCase
 	protected function setUp()
 	{
 		$context = clone \TestHelperMShop::getContext();
-		$context->getConfig()->set( 'mshop/index/manager/text/name', 'MySQL' );
-
 		$this->editor = $context->getEditor();
 		$config = $context->getConfig();
 
-		$dbadapter = $config->get( 'resource/db-product/adapter', $config->get( 'resource/db/adapter' ) );
+		$dbadapter = $config->get( 'resource/db-index/adapter', $config->get( 'resource/db/adapter' ) );
 
 		if( $dbadapter !== 'mysql' ) {
 			$this->markTestSkipped( 'MySQL specific test' );
 		}
 
+		$context->getConfig()->set( 'mshop/index/manager/text/name', 'MySQL' );
 		$this->object = new \Aimeos\MShop\Index\Manager\MySQL( $context );
 	}
